+++ haystack_experimental/components/tools/openapi/_openapi.py
@@ -2,16 +2,14 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-import json
 import logging
 import os
-from dataclasses import dataclass, field
+from collections import defaultdict
 from pathlib import Path
-from typing import Any, Callable, Dict, List, Literal, Optional, Union
+from typing import Any, Callable, Dict, List, Optional, Union
 from urllib.parse import urlparse
 
 import requests
-import yaml
 
 from haystack_experimental.components.tools.openapi._payload_extraction import (
     create_function_payload_extractor,
@@ -21,18 +19,8 @@ from haystack_experimental.components.tools.openapi._schema_conversion import (
     cohere_converter,
     openai_converter,
 )
-from haystack_experimental.components.tools.openapi.types import LLMProvider
+from haystack_experimental.components.tools.openapi.types import LLMProvider, OpenAPISpecification, Operation
 
-VALID_HTTP_METHODS = [
-    "get",
-    "put",
-    "post",
-    "delete",
-    "options",
-    "head",
-    "patch",
-    "trace",
-]
 MIN_REQUIRED_OPENAPI_SPEC_VERSION = 3
 logger = logging.getLogger(__name__)
 
@@ -42,7 +30,7 @@ def is_valid_http_url(url: str) -> bool:
     Check if a URL is a valid HTTP/HTTPS URL.
 
     :param url: The URL to check.
-    :return: True if the URL is a valid HTTP/HTTPS URL, False otherwise.
+    :returns: True if the URL is a valid HTTP/HTTPS URL, False otherwise.
     """
     r = urlparse(url)
     return all([r.scheme in ["http", "https"], r.netloc])
@@ -53,7 +41,7 @@ def send_request(request: Dict[str, Any]) -> Dict[str, Any]:
     Send an HTTP request and return the response.
 
     :param request: The request to send.
-    :return: The response from the server.
+    :returns: The response from the server.
     """
     url = request["url"]
     headers = {**request.get("headers", {})}
@@ -86,7 +74,7 @@ def create_api_key_auth_function(api_key: str):
     Create a function that applies the API key authentication strategy to a given request.
 
     :param api_key: the API key to use for authentication.
-    :return: a function that applies the API key authentication to a request
+    :returns: a function that applies the API key authentication to a request
     at the schema specified location.
     """
 
@@ -117,7 +105,7 @@ def create_http_auth_function(token: str):
     Create a function that applies the http authentication strategy to a given request.
 
     :param token: the authentication token to use.
-    :return: a function that applies the API key authentication to a request
+    :returns: a function that applies the API key authentication to a request
     at the schema specified location.
     """
 
@@ -152,206 +140,6 @@ class HttpClientError(Exception):
     """Exception raised for errors in the HTTP client."""
 
 
-@dataclass
-class Operation:
-    """
-     Represents an operation in an OpenAPI specification
-
-     See https://spec.openapis.org/oas/latest.html#paths-object for details.
-     Path objects can contain multiple operations, each with a unique combination of path and method.
-
-     Attributes:
-        path (str): Path of the operation.
-        method (str): HTTP method of the operation.
-        operation_dict (Dict[str, Any]): Operation details from OpenAPI spec.
-        spec_dict (Dict[str, Any]): The encompassing OpenAPI specification.
-        security_requirements (List[Dict[str, List[str]]]): Security requirements for the operation.
-        request_body (Dict[str, Any]): Request body details.
-        parameters (List[Dict[str, Any]]): Parameters for the operation.
-    """
-
-    path: str
-    method: str
-    operation_dict: Dict[str, Any]
-    spec_dict: Dict[str, Any]
-    security_requirements: List[Dict[str, List[str]]] = field(init=False)
-    request_body: Dict[str, Any] = field(init=False)
-    parameters: List[Dict[str, Any]] = field(init=False)
-
-    def __post_init__(self):
-        if self.method.lower() not in VALID_HTTP_METHODS:
-            raise ValueError(f"Invalid HTTP method: {self.method}")
-        self.method = self.method.lower()
-        self.security_requirements = self.operation_dict.get(
-            "security", []
-        ) or self.spec_dict.get("security", [])
-        self.request_body = self.operation_dict.get("requestBody", {})
-        self.parameters = self.operation_dict.get(
-            "parameters", []
-        ) + self.spec_dict.get("paths", {}).get(self.path, {}).get("parameters", [])
-
-    def get_parameters(
-        self, location: Optional[Literal["header", "query", "path"]] = None
-    ) -> List[Dict[str, Any]]:
-        """
-        Get the parameters for the operation.
-
-        :param location: The location of the parameters to get.
-        :return: The parameters for the operation as a list of dictionaries.
-        """
-        if location:
-            return [param for param in self.parameters if param["in"] == location]
-        return self.parameters
-
-    def get_server(self, server_index: int = 0) -> str:
-        """
-        Get the servers for the operation.
-
-        :param server_index: The index of the server to use.
-        :return: The server URL.
-        :raises ValueError: If no servers are found in the specification.
-        """
-        servers = self.operation_dict.get("servers", []) or self.spec_dict.get(
-            "servers", []
-        )
-        if not servers:
-            raise ValueError("No servers found in the provided specification.")
-        if server_index >= len(servers):
-            raise ValueError(
-                f"Server index {server_index} is out of bounds. "
-                f"Only {len(servers)} servers found."
-            )
-        return servers[server_index].get(
-            "url"
-        )  # just use the first server from the list
-
-
-class OpenAPISpecification:
-    """
-    Represents an OpenAPI specification. See https://spec.openapis.org/oas/latest.html for details.
-    """
-
-    def __init__(self, spec_dict: Dict[str, Any]):
-        if not isinstance(spec_dict, Dict):
-            raise ValueError(
-                f"Invalid OpenAPI specification, expected a dictionary: {spec_dict}"
-            )
-        # just a crude sanity check, by no means a full validation
-        if (
-            "openapi" not in spec_dict
-            or "paths" not in spec_dict
-            or "servers" not in spec_dict
-        ):
-            raise ValueError(
-                "Invalid OpenAPI specification format. See https://swagger.io/specification/ for details.",
-                spec_dict,
-            )
-        self.spec_dict = spec_dict
-
-    @classmethod
-    def from_str(cls, content: str) -> "OpenAPISpecification":
-        """
-        Create an OpenAPISpecification instance from a string.
-
-        :param content: The string content of the OpenAPI specification.
-        :return: The OpenAPISpecification instance.
-        """
-        try:
-            loaded_spec = json.loads(content)
-        except json.JSONDecodeError:
-            try:
-                loaded_spec = yaml.safe_load(content)
-            except yaml.YAMLError as e:
-                raise ValueError(
-                    "Content cannot be decoded as JSON or YAML: " + str(e)
-                ) from e
-        return cls(loaded_spec)
-
-    @classmethod
-    def from_file(cls, spec_file: Union[str, Path]) -> "OpenAPISpecification":
-        """
-        Create an OpenAPISpecification instance from a file.
-
-        :param spec_file: The file path to the OpenAPI specification.
-        :return: The OpenAPISpecification instance.
-        """
-        with open(spec_file, encoding="utf-8") as file:
-            content = file.read()
-        return cls.from_str(content)
-
-    @classmethod
-    def from_url(cls, url: str) -> "OpenAPISpecification":
-        """
-        Create an OpenAPISpecification instance from a URL.
-
-        :param url: The URL to fetch the OpenAPI specification from.
-        :return: The OpenAPISpecification instance.
-        """
-        try:
-            response = requests.get(url, timeout=10)
-            response.raise_for_status()
-            content = response.text
-        except requests.RequestException as e:
-            raise ConnectionError(
-                f"Failed to fetch the specification from URL: {url}. {e!s}"
-            ) from e
-        return cls.from_str(content)
-
-    def find_operation_by_id(
-        self, op_id: str, method: Optional[str] = None
-    ) -> Operation:
-        """
-        Find an Operation by operationId.
-
-        :param op_id: The operationId of the operation.
-        :param method: The HTTP method of the operation.
-        :return: The matching operation
-        :raises ValueError: If no operation is found with the given operationId.
-        """
-        for path, path_item in self.spec_dict.get("paths", {}).items():
-            op: Operation = self.get_operation_item(path, path_item, method)
-            if op_id in op.operation_dict.get("operationId", ""):
-                return self.get_operation_item(path, path_item, method)
-        raise ValueError(
-            f"No operation found with operationId {op_id}, method {method}"
-        )
-
-    def get_operation_item(
-        self, path: str, path_item: Dict[str, Any], method: Optional[str] = None
-    ) -> Operation:
-        """
-        Gets a particular Operation item from the OpenAPI specification given the path and method.
-
-        :param path: The path of the operation.
-        :param path_item: The path item from the OpenAPI specification.
-        :param method: The HTTP method of the operation.
-        :return: The operation
-        """
-        if method:
-            operation_dict = path_item.get(method.lower(), {})
-            if not operation_dict:
-                raise ValueError(
-                    f"No operation found for method {method} at path {path}"
-                )
-            return Operation(path, method.lower(), operation_dict, self.spec_dict)
-        if len(path_item) == 1:
-            method, operation_dict = next(iter(path_item.items()))
-            return Operation(path, method, operation_dict, self.spec_dict)
-        if len(path_item) > 1:
-            raise ValueError(
-                f"Multiple operations found at path {path}, method parameter is required."
-            )
-        raise ValueError(f"No operations found at path {path} and method {method}")
-
-    def get_security_schemes(self) -> Dict[str, Dict[str, Any]]:
-        """
-        Get the security schemes from the OpenAPI specification.
-
-        :return: The security schemes as a dictionary.
-        """
-        return self.spec_dict.get("components", {}).get("securitySchemes", {})
-
-
 class ClientConfiguration:
     """Configuration for the OpenAPI client."""
 
@@ -362,13 +150,22 @@ class ClientConfiguration:
         request_sender: Optional[Callable[[Dict[str, Any]], Dict[str, Any]]] = None,
         llm_provider: Optional[LLMProvider] = None,
     ):  # noqa: PLR0913
+        """
+        Initialize a ClientConfiguration instance.
+
+        :param openapi_spec: The OpenAPI specification as a file path, URL, or dictionary.
+        :param credentials: The credentials to use for authentication.
+        :param request_sender: The function to use for sending requests.
+        :param llm_provider: The LLM provider to use for generating tools definitions.
+        :raises ValueError: If the OpenAPI specification format is invalid.
+        """
         if isinstance(openapi_spec, (str, Path)) and os.path.isfile(openapi_spec):
             self.openapi_spec = OpenAPISpecification.from_file(openapi_spec)
         elif isinstance(openapi_spec, str):
             if is_valid_http_url(openapi_spec):
                 self.openapi_spec = OpenAPISpecification.from_url(openapi_spec)
             else:
-                self.openapi_spec = OpenAPISpecification.from_str(openapi_spec)
+                self.openapi_spec = OpenAPISpecification._from_str(openapi_spec)
         else:
             raise ValueError(
                 "Invalid OpenAPI specification format. Expected file path or dictionary."
@@ -385,7 +182,8 @@ class ClientConfiguration:
         The function takes a security scheme and a request as arguments:
             `security_scheme: Dict[str, Any] - The security scheme from the OpenAPI spec.`
             `request: Dict[str, Any] - The request to apply the authentication to.`
-        :return: The authentication function.
+        :returns: The authentication function.
+        :raises ValueError: If the credentials type is not supported.
         """
         security_schemes = self.openapi_spec.get_security_schemes()
         if not self.credentials:
@@ -400,42 +198,46 @@ class ClientConfiguration:
         """
         Get the tools definitions used as tools LLM parameter.
 
-        :return: The tools definitions passed to the LLM as tools parameter.
+        :returns: The tools definitions passed to the LLM as tools parameter.
         """
-        provider_to_converter = {
-            "anthropic": anthropic_converter,
-            "cohere": cohere_converter,
-        }
-        converter = provider_to_converter.get(self.llm_provider.value, openai_converter)
+        provider_to_converter = defaultdict(
+            lambda: openai_converter,
+            {
+                LLMProvider.ANTHROPIC.value: anthropic_converter,
+                LLMProvider.COHERE.value: cohere_converter,
+            }
+        )
+        converter = provider_to_converter[self.llm_provider.value]
         return converter(self.openapi_spec)
 
-    def get_payload_extractor(self):
+    def get_payload_extractor(self) -> Callable[[Dict[str, Any]], Dict[str, Any]]:
         """
         Get the payload extractor for the LLM provider.
 
         This function knows how to extract the exact function payload from the LLM generated function calling payload.
-        :return: The payload extractor function.
+        :returns: The payload extractor function.
         """
-        provider_to_arguments_field_name = {
-            "anthropic": "input",
-            "cohere": "parameters",
-        }  # add more providers here
-        # default to OpenAI "arguments"
-        arguments_field_name = provider_to_arguments_field_name.get(
-            self.llm_provider.value, "arguments"
+        provider_to_arguments_field_name = defaultdict(
+            lambda: "arguments",
+            {
+                LLMProvider.ANTHROPIC.value: "input",
+                LLMProvider.COHERE.value: "parameters",
+            }
         )
+        arguments_field_name = provider_to_arguments_field_name[self.llm_provider.value]
         return create_function_payload_extractor(arguments_field_name)
 
     def _create_authentication_from_string(
-        self, credentials: str, security_schemes: Dict[str, Any]
+            self, credentials: str, security_schemes: Dict[str, Any]
     ) -> Callable[[Dict[str, Any], Dict[str, Any]], Any]:
         for scheme in security_schemes.values():
             if scheme["type"] == "apiKey":
                 return create_api_key_auth_function(api_key=credentials)
             if scheme["type"] == "http":
                 return create_http_auth_function(token=credentials)
-            if scheme["type"] == "oauth2":
-                raise NotImplementedError("OAuth2 authentication is not yet supported.")
+            raise ValueError(
+                f"Unsupported authentication type '{scheme['type']}' provided."
+            )
         raise ValueError(
             f"Unable to create authentication from provided credentials: {credentials}"
         )
@@ -447,7 +249,9 @@ def build_request(operation: Operation, **kwargs) -> Dict[str, Any]:
 
     :param operation: The operation to build the request for.
     :param kwargs: The arguments to use for building the request.
-    :return: The HTTP request as a dictionary.
+    :returns: The HTTP request as a dictionary.
+    :raises ValueError: If a required parameter is missing.
+    :raises NotImplementedError: If the request body content type is not supported. We only support JSON payloads.
     """
     path = operation.path
     for parameter in operation.get_parameters("path"):
@@ -527,7 +331,6 @@ class OpenAPIServiceClient:
 
     def __init__(self, client_config: ClientConfiguration):
         self.client_config = client_config
-        self.request_sender = client_config.request_sender
 
     def invoke(self, function_payload: Any) -> Any:
         """
@@ -538,19 +341,25 @@ class OpenAPIServiceClient:
         :raises OpenAPIClientError: If the function invocation payload cannot be extracted from the function payload.
         :raises HttpClientError: If an error occurs while sending the request and receiving the response.
         """
-        fn_extractor = self.client_config.get_payload_extractor()
-        fn_invocation_payload = fn_extractor(function_payload)
-        if not fn_invocation_payload:
+        fn_invocation_payload = {}
+        try:
+            fn_extractor = self.client_config.get_payload_extractor()
+            fn_invocation_payload = fn_extractor(function_payload)
+        except Exception as e:
             raise OpenAPIClientError(
-                f"Failed to extract function invocation payload from {function_payload}"
+                f"Error extracting function invocation payload: {str(e)}"
+            ) from e
+
+        if "name" not in fn_invocation_payload or "arguments" not in fn_invocation_payload:
+            raise OpenAPIClientError(
+                f"Function invocation payload does not contain 'name' or 'arguments' keys: {fn_invocation_payload}, "
+                f"the payload extraction function may be incorrect."
             )
         # fn_invocation_payload, if not empty, guaranteed to have "name" and "arguments" keys from here on
-        operation = self.client_config.openapi_spec.find_operation_by_id(
-            fn_invocation_payload.get("name")
-        )
-        request = build_request(operation, **fn_invocation_payload.get("arguments"))
+        operation = self.client_config.openapi_spec.find_operation_by_id(fn_invocation_payload["name"])
+        request = build_request(operation, **fn_invocation_payload["arguments"])
         apply_authentication(self.client_config.get_auth_function(), operation, request)
-        return self.request_sender(request)
+        return self.client_config.request_sender(request)
 
 
 class OpenAPIClientError(Exception):
